+++ backend/src/main/resources/db/migracoes/V23__criacao_tabela_evolucoes_enfermagem.sql
@@ -34,7 +34,9 @@ CREATE TABLE evolucoes_enfermagem (
     exsudato_tipo               TEXT,
     bordas                      VARCHAR(20),
     pele_perilesional           TEXT,
-    dor_escala                  SMALLINT CHECK (dor_escala IS NULL OR (dor_escala BETWEEN 0 AND 10)),
+    -- INTEGER (nao SMALLINT): a entidade mapeia Integer e o Hibernate valida
+    -- o tipo no boot (ddl-auto: validate nos perfis dev e prod).
+    dor_escala                  INTEGER CHECK (dor_escala IS NULL OR (dor_escala BETWEEN 0 AND 10)),
     sinais_vitais               TEXT,
     pa                          VARCHAR(30),
     fc                          VARCHAR(30),
